Remove namespace from form builders & specs
