--- app/form_builders/peoplefinder/form_builder.rb
@@ -1,5 +1,0 @@
-module Peoplefinder
-  class FormBuilder < ActionView::Helpers::FormBuilder
-    attr_reader :template # It's hidden in an instance variable
-  end
-end
