+++ app/form_builders/template_exposing_form_builder.rb
@@ -1,0 +1,3 @@
+class TemplateExposingFormBuilder < ActionView::Helpers::FormBuilder
+  attr_reader :template # It's hidden in an instance variable
+end
